ReferralOnboardingGuideAnimation implemented
