+++ VultisigApp/VultisigApp/Localizables/en.lproj/Localizable.strings
@@ -670,3 +670,4 @@
 "referralOnboardingDescription3" = "Get paid 10bps in your preferred asset every time they trade.";
 "referralProgram" = "Referral Program";
 "howItWorks" = "How it works";
+"getStarted" = "Get Started";
